Run TypeScript tests with swc
