+++ package.json
@@ -1209,6 +1209,12 @@
     "vscode-languageclient": "9.0.1"
   },
   "devDependencies": {
+    "@swc-node/register": "1.6.8",
+    "@swc/core": "1.3.95",
+    "@tsconfig/node20": "20.1.2",
+    "@tsconfig/strictest": "2.0.2",
+    "@types/fs-extra": "11.0.3",
+    "@types/mocha": "10.0.3",
     "@types/vscode": "1.82.0",
     "@vscode/test-electron": "2.3.6",
     "@vscode/vsce": "2.22.0",
@@ -1219,7 +1225,8 @@
     "npm-run-all": "4.1.5",
     "ovsx": "0.8.3",
     "prettier": "3.0.3",
-    "prettier-plugin-packagejson": "2.4.6"
+    "prettier-plugin-packagejson": "2.4.6",
+    "typescript": "5.2.2"
   },
   "packageManager": "yarn@4.0.1",
   "engines": {
