+++ tests/suite/basic/index.js
@@ -2,14 +2,21 @@ const path = require("node:path");
 
 const { globSync } = require("glob");
 const Mocha = require("mocha");
+// @ts-ignore
+const register = require("@swc-node/register");
 
 async function run() {
-  const mocha = new Mocha({ ui: "tdd", color: true, timeout: Infinity });
+  const mocha = new Mocha({
+    ui: "tdd",
+    color: true,
+    require: [register],
+    timeout: Infinity,
+  });
 
   const testsRoot = __dirname;
 
   return new Promise((resolve, reject) => {
-    const files = globSync("**/*.test.js", { cwd: testsRoot });
+    const files = globSync("**/*.test.ts", { cwd: testsRoot });
 
     for (const file of files) {
       mocha.addFile(path.resolve(testsRoot, file));
@@ -20,7 +27,7 @@ async function run() {
         if (failures > 0) {
           reject(new Error(`${failures} tests failed.`));
         } else {
-          resolve();
+          resolve(void 0);
         }
       });
     } catch (err) {
